+++ src/main/java/org/generalka/storage/AnswerDao.java
@@ -11,14 +11,14 @@ public interface AnswerDao {
     void updateAnswer(Answer answer) throws EntityNotFoundException;;
 
     // Delete an answer by its ID
-    void deleteAnswer(int answerId) throws EntityNotFoundException;;
+    void deleteAnswer(Long answerId) throws EntityNotFoundException;;
 
     // Get an answer by its ID
-    Answer getAnswerById(int answerId) throws EntityNotFoundException;;
+    Answer getAnswerById(Long answerId) throws EntityNotFoundException;;
 
     // Get all answers
     List<Answer> getAllAnswers() throws EntityNotFoundException;;
 
-    List<Answer> getAnswersByQuestionId(int questionId) throws EntityNotFoundException;;
+    List<Answer> getAnswersByQuestionId(Long questionId) throws EntityNotFoundException;;
 }
