+++ src/main/java/org/generalka/storage/TestHistoryDao.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 public interface TestHistoryDao {
     void addTestHistory(TestHistory testHistory) throws EntityNotFoundException;;
-    TestHistory getTestHistoryById(int id) throws EntityNotFoundException;;
+    TestHistory getTestHistoryById(Long id) throws EntityNotFoundException;;
     List<TestHistory> getAllTestHistories() throws EntityNotFoundException;;
     void updateTestHistory(TestHistory testHistory) throws EntityNotFoundException;;
-    void deleteTestHistory(int id) throws EntityNotFoundException;;
+    void deleteTestHistory(Long id) throws EntityNotFoundException;;
     // Save a new test history entry
     void saveTestHistory(TestHistory testHistory) throws EntityNotFoundException;;
