+++ src/main/java/org/generalka/storage/TestQuestionDao.java
@@ -9,14 +9,14 @@ public interface TestQuestionDao {
     void updateTestQuestion(TestQuestion testQuestion) throws EntityNotFoundException;;
 
     // Delete a test question by its ID
-    void deleteTestQuestion(int testQuestionId) throws EntityNotFoundException;;
+    void deleteTestQuestion(Long testQuestionId) throws EntityNotFoundException;;
 
     // Get a test question by its ID
-    TestQuestion getTestQuestionById(int testQuestionId) throws EntityNotFoundException;;
+    TestQuestion getTestQuestionById(Long testQuestionId) throws EntityNotFoundException;;
 
     // Get all test questions
     List<TestQuestion> getAllTestQuestions() throws EntityNotFoundException;;
 
-    List<TestQuestion> getTestQuestionsByTestId(int testId) throws EntityNotFoundException;;
+    List<TestQuestion> getTestQuestionsByTestId(Long testId) throws EntityNotFoundException;;
 }
